requirements.txt
pytest
pytest-html
pytest-xdist
pytest-order
pytest-rerunfailures
pytest-base-url
allure-pytest
pyyaml